+++ Runtime/Core/IView.cs
@@ -1,0 +1,10 @@
+namespace Gt.UI
+{
+    public interface IView
+    {
+        bool IsVisible { get; }
+        void Show();
+        void Hide();
+    }
+}
+