+++ Runtime/Presenters/IPresenter.cs
@@ -1,0 +1,9 @@
+namespace Gt.UI
+{
+    public interface IPresenter
+    {
+        void OnShow();
+        void OnHide();
+    }
+}
+